pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "Web_page_Screenshot_Segmentation"
version = "1.0.2"
authors = [
  { name= "Tim Saijun", email = "code@zair.top"},
]
description = "Divide long web page screenshots into blocks to input models with shorter contexts. 将长网页截图进行区块分割，用于输入上下文较短的模型"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["opencv", "website", "image segementation", "screenshot", "gpt"]
dependencies = [
    "opencv-python",
    "numpy",
    "Pillow",
]

[project.urls]
"Homepage" = "https://github.com/Tinnci/Long-Screenshot-Segmentation"
"Documentation" = "https://tim-saijun.github.io/Web_page_Screenshot_Segmentation/"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
screenshot-segment = "Web_page_Screenshot_Segmentation.master:main"
screenshot-draw = "Web_page_Screenshot_Segmentation.drawer:main"
screenshot-split = "Web_page_Screenshot_Segmentation.spliter:main"

[tool.setuptools]
packages = ["Web_page_Screenshot_Segmentation"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["test"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]